Move bob to representations
